refactor: extract config tests from lib.rs into dedicated modules
Split logfwd-config lib.rs (3,406 lines) by extracting inline tests into 7 files:
- tests_config_parsing, tests_validation, tests_input, tests_output,
  tests_enrichment, tests_generator_config, tests_server

lib.rs reduced from 3,406 to 65 lines. All 231 tests pass.
